feat(entity): 更新多个实体类，添加构造器和修改ID字段类型
- 在AlertGroupEntity、ClusterInfoEntity、ClusterServiceCommandEntity、ClusterServiceCommandHostCommandEntity、FrameInfoEntity和UserInfoEntity中添加了无参构造器、全参构造器和SuperBuilder注解，以增强对象创建的灵活性。
- 将ClusterRackEntity中的id字段类型从Integer更改为Long，以确保与数据模型的一致性。

这些改动提升了实体类的可用性和一致性，确保了整体设计的健壮性。

+++ datasophon-dao/src/main/java/com/datasophon/dao/mapper/ClusterRoleUserMapper.java
@@ -61,7 +61,7 @@ public interface ClusterRoleUserMapper extends BaseMapper<ClusterRoleUserEntity>
     /**
      * 根据用户ID和集群ID查询角色用户关系
      */
-    default List<ClusterRoleUserEntity> selectByUserIdAndClusterId(Integer userId, Long clusterId) {
+    default List<ClusterRoleUserEntity> selectByUserIdAndClusterId(Long userId, Long clusterId) {
         QueryWrapper query = QueryWrapper.create()
                 .where(ClusterRoleUserEntity::getUserId).eq(userId)
                 .and(ClusterRoleUserEntity::getClusterId).eq(clusterId);
